Reworked Food 2.0
Removed unnecessary functions from Body.js and Food.js. Added Food.js. Food is now held in a static array. There is a non-static array of snakeTargets, the snake will go after the food in the order it was dropped. Rations will now respawn when the snake runs over them.